backend/fastapi-be/pyproject.toml
[project]
name = "fastapi-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "coloredlogs>=15.0.1",
    "fastapi>=0.115.14",
    "gitpython>=3.1.44",
    "jsonschema>=4.24.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.66",
    "langchain-groq>=0.3.4",
    "langgraph>=0.5.0",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.1",
    "tiktoken>=0.9.0",
    "uvicorn>=0.34.3",
]
